requirements.txt
```
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.23.0
python-dotenv==1.0.0
```